rename io_util.py to util.py
