+++ src/directx/dds.py
@@ -14,7 +14,7 @@ import os
 
 from .dxgi_format import (DXGI_FORMAT, DXGI_BYTE_PER_PIXEL,
                           FOURCC_TO_DXGI, BITMASK_TO_DXGI)
-from io_util import get_size, mkdir
+from util import get_size, mkdir
 
 
 class PF_FLAGS(IntEnum):
